Fix doc build script
