+++ python/pyproject.toml
@@ -36,15 +36,14 @@ numpy = "^1.24.1"
 mypy = "^0.991"
 
 
+
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^2.21.0"
 wheel = "^0.38.4"
 rasterio = {extras = ["s3"], version = "^1.3.4"}
 ipython = "^8.7.0"
 pweave = "^0.30.3"
-jupyter_client = "^5"
-nbclient = "^0.1.0"
-nbconvert = "^5.5.0"
+ipython-genutils = "^0.2.0"
 typer = "^0.7.0"
 pytest = "^4"
